chore(dsp-api-error): use CatalogErrorToResponseTransformer in API and refactor tests
