+++ data-protocols/dsp/dsp-catalog/dsp-catalog-transform/src/main/java/org/eclipse/edc/protocol/dsp/catalog/transform/DspCatalogTransformExtension.java
@@ -53,5 +53,6 @@ public class DspCatalogTransformExtension implements ServiceExtension {
 
         registry.register(new JsonObjectFromCatalogRequestMessageTransformer(jsonFactory, mapper));
         registry.register(new JsonObjectToCatalogRequestMessageTransformer(mapper));
+        registry.register(new CatalogErrorToResponseTransformer());
     }
 }
